fix build and some cleanup

+++ sources/event_manager.cpp
@@ -52,5 +52,5 @@ void EventManager::mouseButtonCallback(GLFWwindow* window, int button, int actio
 void EventManager::cursorPositionCallback(GLFWwindow* window, double xpos, double ypos) {
     Input::cursorPositionCallback(window, xpos, ypos);
 
-    m_events.emplace(Event::MouseMove{xpos, ypos});
+    m_events.emplace(Event::MouseMove{static_cast<float>(xpos), static_cast<float>(ypos)});
 }
